+++ ansible_collections/lvrfrc87/git_acp/plugins/modules/git_acp.py
@@ -15,16 +15,16 @@ DOCUMENTATION = """
 module: git_acp
 author:
     - "Federico Olivieri (@Federico87)"
-short_description: Perform git add, commit and push operations. Set git config user name and email.
+short_description: Perform git add, commit, pull and push operations. Set git config user name and email.
 description:
-    - Manage C(git add), C(git commit) C(git push), C(git config) user name and email on a local
+    - Manage C(git add), C(git commit) C(git push), C(git pull), C(git config) user name and email on a local
       or remote git repository.
 options:
     path:
         description:
             - Folder path where C(.git/) is located.
-        required: true
         type: path
+        required: true
     comment:
         description:
             - Git commit comment. Same as C(git commit -m).
@@ -37,14 +37,6 @@ options:
         type: list
         elements: str
         default: ["."]
-    user:
-        description:
-            - Git username for https operations.
-        type: str
-    token:
-        description:
-            - Git API token for https operations.
-        type: str
     branch:
         description:
             - Git branch where perform git push.
@@ -66,13 +58,6 @@ options:
         description:
             - Git push options. Same as C(git --push-option=option).
         type: str
-    mode:
-        description:
-            - Git operations are performend eithr over ssh, https or local.
-              Same as C(git@git...) or C(https://user:token@git...).
-        choices: ['ssh', 'https', 'local']
-        default: ssh
-        type: str
     url:
         description:
             - Git repo URL.
@@ -91,7 +76,7 @@ options:
                     - If C(yes), ensure that "-o StrictHostKeyChecking=no" is
                       present as an ssh option.
                 type: bool
-                default: 'no'
+                default: False
             ssh_opts:
                 description:
                     - Creates a wrapper script and exports the path as GIT_SSH
@@ -108,20 +93,6 @@ options:
               the normal mechanism for resolving binary paths will be used.
         type: path
         version_added: "1.4.0"
-    remote:
-        description:
-            - Local system alias for git remote PUSH and PULL repository operations.
-        type: str
-        default: origin
-    user_name:
-        description:
-            - Explicit git local user name. Nice to have for remote operations.
-        type: str
-    user_email:
-        description:
-            - Explicit git local email address. Nice to have for remote operations.
-        type: str
-
 requirements:
     - git>=2.10.0 (the command line tool)
 """
@@ -129,41 +100,42 @@ requirements:
 EXAMPLES = """
 - name: HTTPS | add file1.
   git_acp:
-    user: Federico87
-    token: mytoken
-    path: /Users/git/git_acp
-    branch: master
-    comment: Add file1.
-    remote: origin
+    path: "/Users/git/git_acp"
+    comment: "Add file1."
     add: [ "." ]
-    mode: https
-    url: "https://gitlab.com/networkAutomation/git_test_module.git"
+    url: "https://Federico87:mytoken@gitlab.com/networkAutomation/git_test_module.git"
 
-- name: SSH | add file1.
+- name: SSH | add file2.
   git_acp:
-    path: /Users/git/git_acp
-    branch: master
-    comment: Add file1.
-    add: [ file1  ]
-    remote: dev_test
-    mode: ssh
-    url: "git@gitlab.com:networkAutomation/git_test_module.git"
-    user_name: lvrfrc87
-    user_email: lvrfrc87@gmail.com
+    path: "/Users/git/git_acp"
+    branch: development
+    comment: Add file2.
+    add: [ "file2" ]
+    url: "git@gitlab.com:networkAutomation/git_test_module.git dev_test"
 
 - name: LOCAL | push on local repo.
   git_acp:
     path: "~/test_directory/repo"
-    branch: master
-    comment: Add file1.
-    add: [ file1 ]
-    mode: local
+    comment: Add file3.
+    add: [ "file3" ]
     url: /Users/federicoolivieri/test_directory/repo.git
+
+- name: SSH | pull before to push.
+  git_acp:
+    add: [ "c.txt" ]
+    comment: "commit 3"
+    path: "~/test_directory/repo"
+    pull: true
+    url: "git@gitlab.com:networkAutomation/git_test_module.git automation"
+    ssh_params:
+        accept_hostkey: true
+        key_file: "{{ github_ssh_private_key }}"
+        ssh_opts: "-o UserKnownHostsFile={{ remote_tmp_dir }}/known_hosts"
 """
 
 RETURN = """
 output:
-    description: list of git cli command stdout
+    description: list of git cli commands stdout
     type: list
     returned: always
     sample: [
@@ -173,9 +145,6 @@ output:
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.lvrfrc87.git_acp.plugins.module_utils.git_actions import Git
-from ansible_collections.lvrfrc87.git_acp.plugins.module_utils.git_configuration import (
-    GitConfiguration,
-)
 
 
 def main():
@@ -193,88 +162,38 @@ def main():
         executable=dict(default=None, type="path"),
         comment=dict(required=True),
         add=dict(type="list", elements="str", default=["."]),
-        user=dict(),
-        token=dict(no_log=True),
         ssh_params=dict(default=None, type="dict", required=False),
         branch=dict(default="main"),
         pull=dict(default=False, type="bool"),
         pull_options=dict(default=["--no-edit"], type="list", elements="str"),
         push_option=dict(default=None, type="str"),
-        mode=dict(choices=["ssh", "https", "local"], default="ssh"),
-        url=dict(required=True),
-        remote=dict(default="origin"),
-        user_name=dict(),
-        user_email=dict(),
+        url=dict(required=True, no_log=True),
     )
-
-    required_if = [
-        ("mode", "https", ["user", "token"]),
-    ]
-
-    required_together = [
-        ["user_name", "user_email"],
-    ]
 
     module = AnsibleModule(
         argument_spec=argument_spec,
-        required_if=required_if,
-        required_together=required_together,
     )
 
     url = module.params.get("url")
     pull = module.params.get("pull")
-    push_option = module.params.get("push_option")
-    mode = module.params.get("mode")
-    user_name = module.params.get("user_name")
-    user_email = module.params.get("user_email")
     ssh_params = module.params.get("ssh_params")
 
-    # We screenscrape a huge amount of git commands so use C
-    # locale anytime we call run_command()
     module.run_command_environ_update = dict(
         LANG="C.UTF-8", LC_ALL="C.UTF-8", LC_MESSAGES="C.UTF-8", LC_CTYPE="C.UTF-8"
     )
+    result = dict(changed=False)
 
-    if mode == "local":
-        if url.startswith(("https://", "git", "ssh://git")):
-            module.fail_json(msg='SSH or HTTPS mode selected but repo is "local')
-
-        if push_option:
-            module.fail_json(msg='"--push-option" not supported with mode "local"')
-
+    if url.startswith("https://"):
         if ssh_params:
-            module.warn(msg='SSH Parameters will be ignored as mode "local"')
+            module.warn('SSH Parameters will be ignored as "https" in url')
 
-    elif mode == "https":
-        if not url.startswith("https://"):
-            module.fail_json(
-                msg="HTTPS mode selected but url ("
-                + url
-                + ') not starting with "https"'
-            )
-        if ssh_params:
-            module.warn('SSH Parameters will be ignored as mode "https"')
-
-    elif mode == "ssh":
-        if not url.startswith(("git", "ssh://git")):
-            module.fail_json(
-                msg="SSH mode selected but url ("
-                + url
-                + ') not starting with "git" or "ssh://git"'
-            )
-
+    if url.startswith(("git", "ssh://git")):
         if url.startswith("ssh://git@github.com"):
             module.fail_json(
                 msg='GitHub does not support "ssh://" URL. Please remove it from url'
             )
 
-    result = dict(changed=False)
-
     git = Git(module)
-
-    if user_name and user_email:
-        result.update(GitConfiguration(module).user_config())
-
     changed_files = git.status()
 
     if changed_files:
@@ -283,6 +202,7 @@ def main():
         if pull:
             result.update(git.pull())
         result.update(git.push())
+        result["changed"] = True
 
     module.exit_json(**result)